+++ package.json
@@ -43,13 +43,13 @@
     "@fortawesome/free-solid-svg-icons": "^6.7.2",
     "@glimmer/component": "^1.1.2",
     "@glimmer/tracking": "^1.1.2",
+    "@nullvoxpopuli/ember-composable-helpers": "^5.3.2",
     "ember-auto-import": "^2.10.0",
     "ember-basic-dropdown": "^8.5.0",
     "ember-changeset": "^4.1.2",
     "ember-changeset-validations": "^4.1.1",
     "ember-cli-babel": "^8.2.0",
     "ember-cli-htmlbars": "^6.3.0",
-    "@nullvoxpopuli/ember-composable-helpers": "^5.3.2",
     "ember-concurrency": "^4.0.2",
     "ember-engines-router-service": "^0.6.0",
     "ember-flatpickr": "^9.0.2",
@@ -82,12 +82,13 @@
     "ember-cli-clean-css": "3.0.0",
     "ember-cli-dependency-checker": "3.3.3",
     "ember-cli-inject-live-reload": "2.1.0",
-    "ember-cli-mirage": "3.0.4",
     "ember-cli-sass": "11.0.1",
     "ember-cli-sri": "2.1.1",
     "ember-cli-terser": "4.0.2",
     "ember-data": "5.3.9",
     "ember-engines": "0.11.0",
+    "ember-inflector": "^6.0.0",
+    "ember-mirage": "^0.4.3",
     "ember-qunit": "9.0.1",
     "ember-source": "6.1.0",
     "ember-source-channel-url": "3.0.0",
