--- tests/dummy/mirage/serializers/application.js
@@ -1,3 +1,0 @@
-import { JSONAPISerializer } from "miragejs";
-
-export default JSONAPISerializer;
