+++ tests/mirage/serializers/application.js
@@ -1,0 +1,3 @@
+import EmberDataSerializer from "ember-mirage/ember-data-serializer";
+
+export default EmberDataSerializer;
